Aplying cleancode
