+++ Kaizen/kaizen.client/src/components/dashboards/PayrollCostPieChart.vue
@@ -1,4 +1,3 @@
-<!-- src/components/dashboards/PayrollCostChart.vue -->
 <script setup>
 import { ref, onMounted } from 'vue'
 import { Pie } from 'vue-chartjs'
@@ -10,10 +9,8 @@ import {
   Legend
 } from 'chart.js'
 
-/* ─────── REGISTRO DE COMPONENTES CHART.JS ─────── */
 Chart.register(ArcElement, Tooltip, Legend)
 
-/* ─────── DATOS REACTIVOS ─────── */
 const chartData = ref({
   labels: ['Beneficios', 'Deducciones obligatorias', 'Cargas Sociales', 'Salarios'],
   datasets: [{
@@ -23,10 +20,9 @@ const chartData = ref({
   }]
 })
 
-/* ─────── OPCIONES DEL GRÁFICO ─────── */
 const chartOptions = {
   responsive: true,
-  maintainAspectRatio: false,      // altura controlada por CSS
+  maintainAspectRatio: false,
   layout: { padding: { top: 10 } },
   plugins: {
     title: {
@@ -60,13 +56,10 @@ const chartOptions = {
   }
 }
 
-/* ─────── HELPER: CONVIERTE A NÚMERO Y GARANTIZA >0 ─────── */
 const toNumber = v => Number.parseFloat(v) || 0
 const sanitize = arr => arr.every(v => v === 0) ? [1, 1, 1, 1] : arr
 
-/* ─────── OBTENCIÓN DE DATOS ─────── */
 onMounted(async () => {
-  /* 1) Averigua companyPk (localStorage ➜ backend) */
   let companyPk = localStorage.getItem('companyPk')
   if (!companyPk) {
     try {
@@ -74,7 +67,7 @@ onMounted(async () => {
       if (res.ok) {
         const { companyId } = await res.json()
         companyPk = companyId
-        companyPk && localStorage.setItem('companyPk', companyPk)
+        if (companyPk) localStorage.setItem('companyPk', companyPk)
       }
     } catch (err) {
       console.error('No se pudo obtener companyPk del backend', err)
@@ -82,28 +75,23 @@ onMounted(async () => {
   }
   if (!companyPk) return console.warn('Sin companyPk — se aborta el fetch')
 
-  /* 2) Pide breakdown al backend */
   try {
     const { data } = await axios.get(
       '/api/owner-dashboard/payroll-cost-breakdown',
       { params: { companyPk } }
     )
-
-    /* 3) Normaliza y actualiza (⚠️ nueva referencia) */
     const dataset = sanitize([
       toNumber(data.beneficios),
       toNumber(data.obligatorias),
       toNumber(data.cargas),
       toNumber(data.salarios)
     ])
-
     chartData.value = {
       ...chartData.value,
       datasets: [{ ...chartData.value.datasets[0], data: dataset }]
     }
   } catch (err) {
     console.error('Error fetching payroll-cost-breakdown:', err)
-    /* Fallback visual (todos 1) */
     chartData.value = {
       ...chartData.value,
       datasets: [{ ...chartData.value.datasets[0], data: [1, 1, 1, 1] }]
@@ -123,7 +111,6 @@ onMounted(async () => {
   width: 100%;
   max-width: 650px;
   margin: 1rem auto 0;
-  /* Altura fija para evitar canvas colapsado */
   height: 340px;
   display: flex;
   align-items: center;
